fix(ublk): drop unnecessary casts on iod sector fields
Clippy flagged `(iod.start_sector << 9) as u64` and `(iod.nr_sectors << 9) as u32`
as same-type casts. These were hidden until the ublk feature went default-on,
which put `src/ublk.rs::imp` under ci.yml's `cargo clippy --all-targets` lint.

`start_sector` is already u64 and `nr_sectors` is u32, so the shifted result
keeps the same type.

+++ src/ublk.rs
@@ -348,8 +348,8 @@ mod imp {
         loop {
             let iod = *q.get_iod(tag);
             let op = iod.op_flags & 0xff;
-            let off = (iod.start_sector << 9) as u64;
-            let bytes = (iod.nr_sectors << 9) as u32;
+            let off = iod.start_sector << 9;
+            let bytes = iod.nr_sectors << 9;
 
             let buf_slice = buf.as_slice();
             let res = if (bytes as usize) <= buf_slice.len() {
